+++ src/utils/organize/people.spec.ts
@@ -1,42 +1,6 @@
+import { flatOrgs } from './organizations.spec';
 import isEqual from 'lodash.isequal';
-import { nestByParentId } from './organizations';
-import {
-  getConnectedOrganizations,
-  getPersonOrganizations,
-  PersonOrganization,
-} from './people';
-
-const flatOrgs: PersonOrganization[] = [
-  {
-    id: 1,
-    sub_orgs: [],
-    title: 'Root Org',
-  },
-  {
-    id: 2,
-    parent: { id: 1, title: 'Root org' },
-    sub_orgs: [],
-    title: 'First sub-org',
-  },
-  {
-    id: 3,
-    parent: { id: 1, title: 'Root org' },
-    sub_orgs: [],
-    title: 'Second sub-org',
-  },
-  {
-    id: 4,
-    parent: { id: 1, title: 'Root org' },
-    sub_orgs: [],
-    title: 'Third sub-org',
-  },
-  {
-    id: 33,
-    parent: { id: 3, title: 'Second sub-org' },
-    sub_orgs: [],
-    title: 'Child of second sub-org',
-  },
-].map((org) => ({ ...org, is_active: true }));
+import { getConnectedOrganizations, getPersonOrganizations } from './people';
 
 const personConnections = [
   {
@@ -75,16 +39,6 @@ describe('People utils', () => {
       expect(personOrgs.length).toEqual(4);
       expect(connectedOrgs.filter((org) => org.id === 3).length).toEqual(0);
       expect(personOrgs.filter((org) => org.id === 3).length).toEqual(1);
-    });
-  });
-
-  describe('nestByParentId()', () => {
-    it('Returns the correct organization tree from a flat array of organizations', async () => {
-      const orgsTree = nestByParentId(flatOrgs, null);
-      expect(orgsTree.length).toEqual(1);
-      expect(orgsTree[0].title).toEqual(flatOrgs[0].title);
-      expect(orgsTree[0]?.sub_orgs?.length).toEqual(3);
-      expect(orgsTree[0]?.sub_orgs[1].sub_orgs[0].id).toEqual(flatOrgs[4].id);
     });
   });
 });
